requirements.txt
streamlit==1.39.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.3.3
langchain-community==0.3.2
faiss-cpu==1.12.0
ollama==0.2.1
pypdf==4.3.1
gitpython==3.1.43
pyyaml==6.0.2
tqdm==4.66.5
numpy>=1.26.4
python-dotenv==1.0.1